+++ src/qcommon/q_string.hpp
@@ -19,14 +19,14 @@ qboolean Q_isintegral( float f );
 // portable case insensitive compare
 int Q_stricmp(const char *s1, const char *s2);
 int	Q_strncmp(const char *s1, const char *s2, int n);
-int	Q_stricmpn(const char *s1, const char *s2, int n);
+int	Q_stricmpn(const char *s1, const char *s2, size_t n);
 char *Q_strlwr( char *s1 );
 char *Q_strupr( char *s1 );
 char *Q_strrchr( const char* string, int c );
 
 // buffer size safe library replacements
 void Q_strncpyz( char *dest, const char *src, size_t destsize );
-void Q_strcat( char *dest, int size, const char *src );
+void Q_strcat( char *dest, size_t size, const char *src );
 
 const char *Q_stristr( const char *s, const char *find);
